Sửa phần User, Role, Permission

+++ service-common/src/main/java/com/global/servicebase/user/model/User.java
@@ -1,12 +1,11 @@
 package com.global.servicebase.user.model;
 
 import com.global.jdbc.SsdcEntity;
-import java.math.BigInteger;
 import org.mindrot.jbcrypt.BCrypt;
 import java.util.LinkedHashSet;
 import java.util.Set;
 
-public class User extends SsdcEntity<BigInteger> {
+public class User extends SsdcEntity<Long> {
 
     public String userName;
     public String fullName;
